+++ src/main/java/pl/lcc/listener/example/security/SecurityConfiguration.java
@@ -4,22 +4,23 @@
  */
 package pl.lcc.listener.example.security;
 
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.http.HttpMethod;
+import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
 import org.springframework.security.crypto.password.NoOpPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.security.web.SecurityFilterChain;
 import org.springframework.security.web.access.channel.ChannelProcessingFilter;
-import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
 
 /**
  *
  * @author Nauczyciel
  */
+@Slf4j
 @Configuration
 @EnableWebSecurity
 public class SecurityConfiguration {
@@ -29,40 +30,40 @@ public class SecurityConfiguration {
 
     @Autowired
     LoggingLogoutHandler loggingLogoutHandler;
-    
-    
+
+    @Autowired
+    UserManegementService umService;
+
     @Bean
     public SecurityFilterChain securityFilterChain2(HttpSecurity http) throws Exception {
 
         http
                 .authorizeRequests()
-                    .antMatchers("/login").permitAll()
-                    .antMatchers("/create").permitAll()
-                    .mvcMatchers("/mod").hasAuthority(Authority.MOD.getAuthority())
-                    .mvcMatchers("/verified").hasAuthority(Authority.MOD.getAuthority())
-                    .anyRequest().authenticated()
-                    .and()
+                .antMatchers("/login").permitAll()
+                .antMatchers("/create").permitAll()
+                .mvcMatchers("/mod").hasAuthority(Authority.MOD.getAuthority())
+                .mvcMatchers("/verified").hasAuthority(Authority.MOD.getAuthority())
+                .anyRequest().authenticated()
+                .and()
                 .formLogin()
-                    .loginPage("/login")
-                    .failureHandler(loginFailureHandler)
-                    .and()
+                .loginPage("/login")
+                .failureHandler(loginFailureHandler)
+                .and()
                 .logout()
-                    .invalidateHttpSession(true)
-                    .deleteCookies("JSESSIONID")
-                    .logoutUrl("/logout")
-                    .logoutSuccessUrl("/login?logout")
-                    .addLogoutHandler(loggingLogoutHandler)
-                    .permitAll()
-                    .and()
+                .invalidateHttpSession(true)
+                .deleteCookies("JSESSIONID")
+                .logoutUrl("/logout")
+                .logoutSuccessUrl("/login?logout")
+                .addLogoutHandler(loggingLogoutHandler)
+                .permitAll()
+                .and()
                 .exceptionHandling()
-                    .accessDeniedPage("/403")
-                    .and()
+                .accessDeniedPage("/403")
+                .and()
                 .sessionManagement()
-                    .invalidSessionUrl("/login?expired")
+                .invalidSessionUrl("/login?expired")
                 .and()
                 .addFilterBefore(new LoggingFilter(), ChannelProcessingFilter.class);
-               
-                    
 
         return http.build();
     }
